+++ deno/payloads/v10/invite.ts
@@ -6,6 +6,7 @@ import type { APIApplication } from './application.ts';
 import type { APIInviteChannel } from './channel.ts';
 import type { APIGuild } from './guild.ts';
 import type { APIGuildScheduledEvent } from './guildScheduledEvent.ts';
+import type { APIRole } from './permissions.ts';
 import type { APIUser } from './user.ts';
 
 export type APIInviteGuild = Pick<
@@ -21,6 +22,10 @@ export type APIInviteGuild = Pick<
 	| 'splash'
 	| 'vanity_url_code'
 	| 'verification_level'
+>;
+
+export type APIInviteRole = Required<
+	Pick<APIRole, 'color' | 'colors' | 'icon' | 'id' | 'name' | 'position' | 'unicode_emoji'>
 >;
 
 /**
@@ -85,7 +90,7 @@ export interface APIInvite {
 	 * @deprecated This has been removed from the documentation.
 	 * {@link https://github.com/discord/discord-api-docs/pull/7779 | discord-api-docs#7779}
 	 */
-	stage_instance?: never;
+	stage_instance?: undefined;
 	/**
 	 * The guild scheduled event data, returned from the `GET /invites/<code>` endpoint when `guild_scheduled_event_id` is a valid guild scheduled event id
 	 */
@@ -98,6 +103,10 @@ export interface APIInvite {
 	 * The flags of the invite
 	 */
 	flags?: InviteFlags;
+	/**
+	 * The roles assigned to the user upon accepting the invite
+	 */
+	roles?: APIInviteRole[];
 }
 
 /**