+++ rest/common.ts
@@ -60,6 +60,10 @@ export enum RESTJSONErrorCodes {
 
 	UnknownSound = 10_097,
 
+	UnknownInviteTargetUsersJob = 10_124,
+
+	UnknownInviteTargetUsers = 10_129,
+
 	BotsCannotUseThisEndpoint = 20_001,
 	OnlyBotsCanUseThisEndpoint,
 
@@ -67,7 +71,8 @@ export enum RESTJSONErrorCodes {
 
 	NotAuthorizedToPerformThisActionOnThisApplication = 20_012,
 
-	ActionCannotBePerformedDueToSlowmodeRateLimit = 20_016,
+	ThisActionRequiresAPremiumSubscription = 20_015,
+	ActionCannotBePerformedDueToSlowmodeRateLimit,
 	TheMazeIsntMeantForYou,
 	OnlyTheOwnerOfThisAccountCanPerformThisAction,
 
@@ -275,6 +280,8 @@ export enum RESTJSONErrorCodes {
 	VoiceMessagesCannotBeEdited,
 	CannotDeleteGuildSubscriptionIntegration,
 
+	CannotSendVoiceEffectWhenUserIsServerMutedDeafenedOrSuppressed = 50_167,
+
 	YouCannotSendVoiceMessagesInThisChannel = 50_173,
 
 	TheUserAccountMustFirstBeVerified = 50_178,
@@ -335,6 +342,7 @@ export enum RESTJSONErrorCodes {
 	WebhookServicesCannotBeUsedInForumChannels,
 
 	MessageBlockedByHarmfulLinksFilter = 240_000,
+	AccessToJoiningNewServersHasBeenLimitedForTheUser = 340_015,
 
 	CannotEnableOnboardingRequirementsAreNotMet = 350_000,
 	CannotUpdateOnboardingWhileBelowRequirements,